requirements.txt
flask==3.1.0
flask-cors==5.0.1
pyjwt==2.10.1
bcrypt==4.3.0
python-dotenv==1.1.0
werkzeug==3.1.3
scikit-surprise==1.1.4
numpy>=1.24.0
pandas>=2.0.0
